requirements.txt
bson==0.5.8
pymongo==3.9.0
bottle==0.12.17
python-dotenv==0.10.3
sklearn==0.0
pandas==0.25.2
scikit-learn==0.21.3
nltk==3.4.5
flair==0.4.4
dnspython==1.16.0